reddit-api/requirements.txt
praw
kafka-python
pyspark
python-dotenv